refactor: api -> handler is now using modern hierarchy style
