+++ src/lib.rs
@@ -1,4 +1,4 @@
-pub mod api;
+pub mod handler;
 pub mod db;
 pub mod model;
 pub mod macros;
